refactor: Retire the models barrel and type the database-write wrapper
Both were left out of the previous pass: the barrel because six of its
importers were on the unmerged #20 branch, and the wrapper because it
was not obviously small until the shape became clear.

models/__init__.py re-exported twenty-two names, so `from models import
Stream` landed a reader on a list rather than on the file that defines
it. services/__init__.py and db/__init__.py were reduced to docstrings
for exactly this reason; models/ was the one left, and unlike
db/models/__init__.py it has no metadata to register, so nothing was
holding it up. Eleven consumers now import from the owning module. The
same rule one level down retires models/twitch_event_subs/common, whose
single consumer is the payload beside it.

_safe_db_operation took a function and `*args: object`, so
upsert_message's six arguments were checked against nothing - they could
be reordered and pyright would agree. It now takes the call already
made, which is the smallest change that restores checking: the call site
writes repository.upsert_message(...) as an ordinary call. Confirmed by
swapping two of those arguments and watching pyright raise two errors
where it previously raised none.

Deliberately still not done, and neither is small: the Stream/LiveAlert
representation split (one value, four representations - #33 names it out
of scope for the same reason), and cogs/tasks.py's second held-back
mechanism, whose "recovered" half is something errors.py cannot express
and so needs a decision rather than a refactor.

Four checks clean in order; gate WARN, non-blocking, on services/audit.py
approaching the 400-line limit - pre-existing, and slightly smaller after
the previous commit removed its redundant channel alias. Verified by
importing all 24 modules and re-running the three harnesses. No test
suite, so this is not tested.
