+++ cogs/events.py
@@ -1,4 +1,4 @@
-from collections.abc import Awaitable, Callable
+from collections.abc import Awaitable
 
 import discord
 import pendulum
@@ -31,15 +31,17 @@ class Events(Cog):
         self.bot = bot
 
     async def _safe_db_operation(
-        self,
-        operation: str,
-        func: Callable[..., Awaitable[object]],
-        *args: object,
-        **kwargs: object,
+        self, operation: str, write: Awaitable[object]
     ) -> None:
-        """Run a database write, reporting failures instead of raising."""
+        """Run a database write, reporting failures instead of raising.
+
+        Takes the call already made rather than a function and its arguments.
+        Passed as `*args: object` they were checked against nothing, so
+        upsert_message's six could be reordered and pyright would agree; written
+        out at the call site they are checked as ordinary arguments.
+        """
         try:
-            await func(*args, **kwargs)
+            await write
         except Exception as e:  # noqa: BLE001
             await report(e, f"Failed to {operation}")
 
@@ -60,13 +62,14 @@ class Events(Cog):
         guild = message.guild
         await self._safe_db_operation(
             f"store message {message.id}",
-            repository.upsert_message,
-            message.id,
-            message.content,
-            config.setting("guild_id") if guild is None else guild.id,
-            message.author.id,
-            message.channel.id,
-            [attachment.url for attachment in message.attachments],
+            repository.upsert_message(
+                message.id,
+                message.content,
+                config.setting("guild_id") if guild is None else guild.id,
+                message.author.id,
+                message.channel.id,
+                [attachment.url for attachment in message.attachments],
+            ),
         )
 
     @Cog.listener()
@@ -108,9 +111,7 @@ class Events(Cog):
 
             await self._safe_db_operation(
                 f"insert user {member.name} ({member.id})",
-                repository.upsert_username,
-                member.id,
-                member.name,
+                repository.upsert_username(member.id, member.name),
             )
         except Exception as e:  # noqa: BLE001
             await report(e, "Fatal error with on_member_join event")
@@ -135,8 +136,7 @@ class Events(Cog):
 
             await self._safe_db_operation(
                 f"remove user {member.name} ({member.id})",
-                repository.delete_user,
-                member.id,
+                repository.delete_user(member.id),
             )
         except Exception as e:  # noqa: BLE001
             await report(e, "Fatal error with on_raw_member_remove event")
@@ -262,8 +262,7 @@ class Events(Cog):
 
             await self._safe_db_operation(
                 f"delete message {payload.message_id}",
-                repository.delete_message,
-                payload.message_id,
+                repository.delete_message(payload.message_id),
             )
         except Exception as e:  # noqa: BLE001
             await report(e, "Fatal error with on_raw_message_delete event")
@@ -286,8 +285,7 @@ class Events(Cog):
             for message_id in payload.message_ids:
                 await self._safe_db_operation(
                     f"delete message {message_id}",
-                    repository.delete_message,
-                    message_id,
+                    repository.delete_message(message_id),
                 )
         except Exception as e:  # noqa: BLE001
             await report(e, "Fatal error with on_raw_bulk_message_delete event")
