+++ services/twitch/token_manager.py
@@ -14,7 +14,7 @@ from constants import TokenType
 from db.models import OAuthToken
 from db.session import session_scope
 from errors import notify
-from models import AuthResponse, RefreshResponse
+from models.auth.auth_response import AuthResponse, RefreshResponse
 from services.http_client import client
 from services.twitch.oauth import configured_scopes
